+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.2.0 (2017-01-18)
+  * [#13](https://github.com/etehtsea/oxblood/issues/13):
+    Added support for ZRANGEBYLEX/ZREMRANGEBYLEX/ZREVRANGEBYLEX - [@etehtsea](https://github.com/etehtsea)
   * [#14](https://github.com/etehtsea/oxblood/issues/14):
-    Added for ZINTERSTORE/ZUNIONSTORE - [@etehtsea](https://github.com/etehtsea)
+    Added support for ZINTERSTORE/ZUNIONSTORE - [@etehtsea](https://github.com/etehtsea)
   * [#12](https://github.com/etehtsea/oxblood/issues/12):
     Added support for ZADD options - [@etehtsea](https://github.com/etehtsea)
   * [#3](https://github.com/etehtsea/oxblood/issues/3),